Create vcp_approx_method.java
Java implementation of vertex cover problem by approximate method. According to the algorithm, we select any arbitrary edge from the graph, then check if both the vertices of the edge are already in the vertex cover set. If both of them are not present in the set, we add them to the set and remove all the edges in the graph that are incident on either of the vertices.